including X-RAY file remover

+++ processing/main.sh
@@ -4,9 +4,13 @@
 
 # get pdb files
 # must remove XRAY based pdb file
-for file in `grep "EXPERIMENT TYPE" * | grep -v "X-RAY" | cut -d: -f1`; do 
-  echo $file; 
-  rm -f $file; 
+for file in `grep "EXPERIMENT TYPE" * | cut -d: -f1`; do   
+    if [ "$file" != "main.sh" ]; then
+        if [ "`grep "X-RAY" $file | cut -d: -f1`" ]; then
+            echo $file; 
+            rm -f $file;
+        fi
+    fi
 done
 
 # call torance
